Simplify the API integration
