+++ app/Http/Controllers/Twill/ArtworkController.php
@@ -3,7 +3,6 @@
 namespace App\Http\Controllers\Twill;
 
 use A17\Twill\Http\Controllers\Admin\ModuleController;
-use A17\Twill\Models\Behaviors\HasTranslation;
 use A17\Twill\Models\Contracts\TwillModelContract;
 use A17\Twill\Services\Forms\BladePartial;
 use A17\Twill\Services\Forms\Fields\Checkbox;
@@ -11,12 +10,8 @@ use A17\Twill\Services\Forms\Fields\Input;
 use A17\Twill\Services\Forms\Fields\Medias;
 use A17\Twill\Services\Forms\Fieldset;
 use A17\Twill\Services\Forms\Form;
-use A17\Twill\Services\Listings\Columns\Text;
-use A17\Twill\Services\Listings\TableColumns;
 use App\Libraries\Api\Builders\ApiQueryBuilder;
 use App\Libraries\Api\Builders\Connection\AicConnection;
-use App\Libraries\Api\Models\Behaviors\HasApiCalls;
-use App\Models\Api\Artwork as ApiArtwork;
 use App\Support\Forms\Fields\QueryArtwork;
 use Exception;
 use Facades\App\Libraries\DamsImageService;
@@ -25,23 +20,10 @@ use Illuminate\Http\Request;
 
 class ArtworkController extends ModuleController
 {
-    use HasApiCalls;
-
     protected $moduleName = 'artworks';
 
     protected function setUpController(): void
     {
-        // $this->disableBulkDelete();
-        // $this->disableBulkEdit();
-        // $this->disableBulkPublish();
-        // $this->disableCreate();
-        // $this->disableDelete();
-        // $this->disableEdit();
-        // $this->disablePermalink();
-        // $this->disablePublish();
-        // $this->disableRestore();
-
-        // $this->enableAugmentedModel();
         $this->enableShowImage();
     }
 
@@ -65,10 +47,8 @@ class ArtworkController extends ModuleController
 
     public function getForm(TwillModelContract $model): Form
     {
-        $model->refreshApi();
-
         $content = Form::make()
-            ->merge($this->additionalFormFields($model, $model->getApiModel()));
+            ->merge($this->additionalFormFields($model));
 
         return parent::getForm($model)
             ->addFieldset(
@@ -79,32 +59,13 @@ class ArtworkController extends ModuleController
             );
     }
 
-    // protected function additionalIndexTableColumns(): TableColumns
-    // {
-    //     $table = parent::additionalIndexTableColumns();
-
-    //     $table->add(
-    //         Text::make()->field('description')->title('Description')
-    //     );
-
-    //     return $table;
-    // }
-
-    protected function additionalFormFields($object, $apiCollectionObject): Form
+    protected function additionalFormFields($object): Form
     {
-        $apiValues = array_map(
-            fn ($value) => $value ?? (string) $value,
-            $apiCollectionObject->getAttributes()
-        );
-
-        $latitude = $object->latitude ?? $apiCollectionObject->latitude ?? '';
-        $longitude = $object->longitude ?? $apiCollectionObject->longitude ?? '';
-
         return Form::make()
             ->add(
                 BladePartial::make()
                     ->view('forms.image')
-                    ->withAdditionalParams(['src' => $apiCollectionObject->imageFront('iiif')['src'] ?? ''])
+                    ->withAdditionalParams(['src' => null])
             )
             ->add(
                 Medias::make()
@@ -115,43 +76,36 @@ class ArtworkController extends ModuleController
             ->add(
                 Input::make()
                     ->name('artist_display')
-                    ->placeholder($apiValues['artist_display'])
             )
             ->add(
                 Checkbox::make()
                     ->name('is_on_view')
-                    ->default($apiValues['is_on_view'])
             )
             ->add(
                 Input::make()
                     ->name('main_reference_number')
-                    ->placeholder($apiValues['main_reference_number'])
                     ->disabled()
                     ->note('readonly')
             )
             ->add(
                 Input::make()
                     ->name('credit_line')
-                    ->placeholder($apiValues['credit_line'])
             )
             ->add(
                 Input::make()
                     ->name('copyright_notice')
-                    ->placeholder($apiValues['copyright_notice'])
             )
             ->add(
                 Input::make()
                     ->name('latitude')
                     ->label('Latitude')
                     ->type('number')
-                    ->placeholder($latitude)
             )
             ->add(
                 Input::make()
                     ->name('longitude')
                     ->label('Longitude')
                     ->type('number')
-                    ->placeholder($longitude)
             );
     }
