--- app/Libraries/Api/Models/BaseApiModel.php
@@ -1,941 +1,0 @@
-<?php
-
-/**
- * Pretty much a port from an Eloquent model to a class without a backend
- * Based/copied from https://github.com/jenssegers/model
- */
-
-namespace App\Libraries\Api\Models;
-
-use A17\Twill\Facades\TwillPermissions;
-use A17\Twill\Models\Behaviors\HasPresenter;
-use A17\Twill\Models\Behaviors\IsTranslatable;
-use A17\Twill\Models\Contracts\TwillModelContract;
-use A17\Twill\Models\Permission;
-use App\Libraries\Api\Models\ApiCollection as BaseCollection;
-use App\Libraries\Api\Models\Behaviors\HasApiCalls;
-use App\Libraries\Api\Models\Behaviors\HasAugmentedModel;
-use ArrayAccess;
-use DateTimeInterface;
-use Illuminate\Contracts\Routing\UrlRoutable;
-use Illuminate\Contracts\Support\Arrayable;
-use Illuminate\Contracts\Support\Jsonable;
-use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\MassAssignmentException;
-use Illuminate\Support\Arr;
-use Illuminate\Support\Carbon;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Str;
-use JsonSerializable;
-use Stringable;
-
-abstract class BaseApiModel implements Arrayable, ArrayAccess, Jsonable, JsonSerializable, Stringable, TwillModelContract, UrlRoutable
-{
-    use HasApiCalls;
-    use HasAugmentedModel;
-    use HasPresenter;
-    use IsTranslatable;
-
-    protected array $attributes = [];
-
-    /**
-     * The attributes that should be hidden for arrays.
-     */
-    protected array $hidden = [];
-
-    /**
-     * The attributes that should be visible in arrays.
-     */
-    protected array $visible = [];
-
-    /**
-     * The accessors to append to the model's array form.
-     */
-    protected array $appends = [];
-
-    /**
-     * The attributes that are mass assignable.
-     */
-    protected array $fillable = [];
-
-    /**
-     * The attributes that aren't mass assignable.
-     */
-    protected array $guarded = [];
-
-    /**
-     * The attributes that should be casted to native types.
-     */
-    protected array $casts = [];
-
-    /**
-     * Used to create URL's
-     */
-    protected string $primaryKey = 'id';
-
-    /**
-     * Indicates whether attributes are snake cased on arrays.
-     */
-    public static bool $snakeAttributes = true;
-
-    /**
-     * Indicates if all mass assignment is enabled.
-     */
-    protected static bool $unguarded = false;
-
-    /**
-     * The cache of the mutated attributes for each class.
-     */
-    protected static array $mutatorCache = [];
-
-    /**
-     * The relations to eager load on every query.
-     */
-    protected array $with = [];
-
-    /**
-     * Endpoints for this entity
-     */
-    protected array $endpoints = [
-        'collection' => '',
-        'resource' => '',
-        'search' => '',
-    ];
-
-    /**
-     * Create a new Eloquent model instance.
-     */
-    public function __construct(array $attributes = [])
-    {
-        $this->fill($attributes);
-    }
-
-    public function scopePublished(Builder $query): Builder
-    {
-        if (config('aic.is_preview_mode')) {
-            return $query;
-        }
-
-        // scopeVisible checks for its fillable columns
-        $query->visible();
-
-        if ($this->isFillable('published')) {
-            $query->wherePublished(true);
-        }
-
-        return $query;
-    }
-
-    public function scopeAccessible(Builder $query): Builder
-    {
-        if (! TwillPermissions::enabled()) {
-            return $query;
-        }
-
-        $model = $query->getModel()::class;
-        $moduleName = TwillPermissions::getPermissionModule(getModuleNameByModel($model));
-
-        if ($moduleName && ! Auth::user()->isSuperAdmin()) {
-            // Get all permissions the logged in user has regards to the model.
-            $allPermissions = Auth::user()->allPermissions();
-            $allModelPermissions = (clone $allPermissions)->ofModel($model);
-
-            // If the user has any module permissions, or global manage all modules permissions, all items will be return
-            if (
-                (clone $allModelPermissions)->module()
-                    ->whereIn('name', Permission::available(Permission::SCOPE_MODULE))
-                    ->exists()
-                || (clone $allPermissions)->global()->where('name', 'manage-modules')->exists()
-            ) {
-                return $query;
-            }
-
-            // If the module is submodule, skip the scope.
-            if (strpos((string) $moduleName, '.')) {
-                return $query;
-            }
-
-            $authorizedItemsIds = $allModelPermissions->moduleItem()->pluck('permissionable_id');
-
-            return $query->whereIn('id', $authorizedItemsIds);
-        }
-
-        return $query;
-    }
-
-    public function scopeOnlyTrashed(Builder $query): Builder
-    {
-        return $query->onlyTrashed();
-    }
-
-    public function scopeDraft(Builder $query): Builder
-    {
-        return $query->where('published', false);
-    }
-
-    public function getTranslatedAttributes(): array
-    {
-        return $this->translatedAttributes ?? [];
-    }
-
-    /**
-     * Fill the model with an array of attributes.
-     *
-     * @throws MassAssignmentException
-     */
-    public function fill(array $attributes): static
-    {
-        $totallyGuarded = $this->totallyGuarded();
-
-        foreach ($this->fillableFromArray($attributes) as $key => $value) {
-            // The developers may choose to place some attributes in the "fillable"
-            // array, which means only those attributes may be set through mass
-            // assignment to the model, and all others will just be ignored.
-            if ($this->isFillable($key)) {
-                $this->setAttribute($key, $value);
-            } elseif ($totallyGuarded) {
-                throw new MassAssignmentException($key);
-            }
-        }
-
-        return $this;
-    }
-
-    /**
-     * Fill the model with an array of attributes. Force mass assignment.
-     */
-    public function forceFill(array $attributes): static
-    {
-        // Since some versions of PHP have a bug that prevents it from properly
-        // binding the late static context in a closure, we will first store
-        // the model in a variable, which we will then use in the closure.
-        $model = $this;
-
-        return static::unguarded(fn () => $model->fill($attributes));
-    }
-
-    /**
-     * Get the fillable attributes of a given array.
-     */
-    protected function fillableFromArray(array $attributes): array
-    {
-        if ($this->fillable !== [] && ! static::$unguarded) {
-            return array_intersect_key($attributes, array_flip($this->fillable));
-        }
-
-        return $attributes;
-    }
-
-    /**
-     * Create a new instance of the given model.
-     */
-    public function newInstance($attributes = []): static
-    {
-        return new static((array) $attributes);
-    }
-
-    /**
-     * Create a collection of models from plain arrays.
-     */
-    public static function hydrate(array $items): array
-    {
-        $instance = new static();
-
-        return array_map(fn ($item) => $instance->newInstance($item), $items);
-    }
-
-    /**
-     * Get the hidden attributes for the model.
-     */
-    public function getHidden(): array
-    {
-        return $this->hidden;
-    }
-
-    /**
-     * Set the hidden attributes for the model.
-     */
-    public function setHidden(array $hidden): static
-    {
-        $this->hidden = $hidden;
-
-        return $this;
-    }
-
-    /**
-     * Make the given, typically visible, attributes hidden.
-     */
-    public function makeHidden(mixed $attributes = null): static
-    {
-        $attributes = is_array($attributes) ? $attributes : func_get_args();
-
-        $this->hidden = array_merge($this->hidden, $attributes);
-
-        return $this;
-    }
-
-    /**
-     * Make the given, typically hidden, attributes visible.
-     */
-    public function withHidden(mixed $attributes): static
-    {
-        $this->hidden = array_diff($this->hidden, (array) $attributes);
-
-        return $this;
-    }
-
-    /**
-     * Get the visible attributes for the model.
-     */
-    public function getVisible(): array
-    {
-        return $this->visible;
-    }
-
-    /**
-     * Set the visible attributes for the model.
-     */
-    public function setVisible(array $visible): static
-    {
-        $this->visible = $visible;
-
-        return $this;
-    }
-
-    /**
-     * Make the given, typically hidden, attributes visible.
-     */
-    public function makeVisible(mixed $attributes = null): static
-    {
-        $attributes = is_array($attributes) ? $attributes : func_get_args();
-
-        $this->visible = array_merge($this->visible, $attributes);
-
-        return $this;
-    }
-
-    /**
-     * Set the accessors to append to model arrays.
-     */
-    public function setAppends(array $appends): static
-    {
-        $this->appends = $appends;
-
-        return $this;
-    }
-
-    /**
-     * Get the fillable attributes for the model.
-     */
-    public function getFillable(): array
-    {
-        return $this->fillable;
-    }
-
-    /**
-     * Set the fillable attributes for the model.
-     */
-    public function fillable(array $fillable): static
-    {
-        $this->fillable = $fillable;
-
-        return $this;
-    }
-
-    /**
-     * Get the guarded attributes for the model.
-     */
-    public function getGuarded(): array
-    {
-        return $this->guarded;
-    }
-
-    /**
-     * Set the guarded attributes for the model.
-     */
-    public function guard(array $guarded): static
-    {
-        $this->guarded = $guarded;
-
-        return $this;
-    }
-
-    /**
-     * Disable all mass assignable restrictions.
-     */
-    public static function unguard(bool $state = true): void
-    {
-        static::$unguarded = $state;
-    }
-
-    /**
-     * Enable the mass assignment restrictions.
-     */
-    public static function reguard(): void
-    {
-        static::$unguarded = false;
-    }
-
-    /**
-     * Determine if current state is "unguarded".
-     */
-    public static function isUnguarded(): bool
-    {
-        return static::$unguarded;
-    }
-
-    /**
-     * Run the given callable while being unguarded.
-     */
-    public static function unguarded(callable $callback): mixed
-    {
-        if (static::$unguarded) {
-            return $callback();
-        }
-
-        static::unguard();
-
-        $result = $callback();
-
-        static::reguard();
-
-        return $result;
-    }
-
-    /**
-     * Determine if the given attribute may be mass assigned.
-     */
-    public function isFillable(string $key): bool
-    {
-        return match (true) {
-            static::$unguarded => true,
-            // If the key is in the "fillable" array, we can of course assume that it's
-            // a fillable attribute. Otherwise, we will check the guarded array when
-            // we need to determine if the attribute is black-listed on the model.
-            in_array($key, $this->fillable) => true,
-            $this->isGuarded($key) => false,
-            default => $this->fillable === [],
-        };
-    }
-
-    /**
-     * Determine if the given key is guarded.
-     */
-    public function isGuarded(string $key): bool
-    {
-        return in_array($key, $this->guarded) || $this->guarded == ['*'];
-    }
-
-    /**
-     * Determine if the model is totally guarded.
-     */
-    public function totallyGuarded(): bool
-    {
-        return count($this->fillable) == 0 && $this->guarded == ['*'];
-    }
-
-    /**
-     * Convert the model instance to JSON.
-     */
-    public function toJson($options = 0): string
-    {
-        return json_encode($this->jsonSerialize(), $options);
-    }
-
-    /**
-     * Convert the object into something JSON serializable.
-     */
-    public function jsonSerialize(): array
-    {
-        return $this->toArray();
-    }
-
-    /**
-     * Convert the model instance to an array.
-     */
-    public function toArray(): array
-    {
-        return $this->attributesToArray();
-    }
-
-    /**
-     * Convert the model's attributes to an array.
-     */
-    public function attributesToArray(): array
-    {
-        $attributes = $this->getArrayableAttributes();
-
-        $mutatedAttributes = $this->getMutatedAttributes();
-
-        // We want to spin through all the mutated attributes for this model and call
-        // the mutator for the attribute. We cache off every mutated attributes so
-        // we don't have to constantly check on attributes that actually change.
-        foreach ($mutatedAttributes as $key) {
-            if (! array_key_exists($key, $attributes)) {
-                continue;
-            }
-
-            $attributes[$key] = $this->mutateAttributeForArray(
-                $key,
-                $attributes[$key]
-            );
-        }
-
-        // Next we will handle any casts that have been setup for this model and cast
-        // the values to their appropriate type. If the attribute has a mutator we
-        // will not perform the cast on those attributes to avoid any confusion.
-        foreach (array_keys($this->casts) as $key) {
-            if (
-                ! array_key_exists($key, $attributes) ||
-                in_array($key, $mutatedAttributes)
-            ) {
-                continue;
-            }
-
-            $attributes[$key] = $this->castAttribute(
-                $key,
-                $attributes[$key]
-            );
-        }
-
-        // Here we will grab all of the appended, calculated attributes to this model
-        // as these attributes are not really in the attributes array, but are run
-        // when we need to array or JSON the model for convenience to the coder.
-        foreach ($this->getArrayableAppends() as $key) {
-            $attributes[$key] = $this->mutateAttributeForArray($key, null);
-        }
-
-        return $attributes;
-    }
-
-    /**
-     * Get an attribute array of all arrayable attributes.
-     */
-    protected function getArrayableAttributes(): array
-    {
-        return $this->getArrayableItems($this->attributes);
-    }
-
-    /**
-     * Get all of the appendable values that are arrayable.
-     */
-    protected function getArrayableAppends(): array
-    {
-        if ($this->appends === []) {
-            return [];
-        }
-
-        return $this->getArrayableItems(
-            array_combine($this->appends, $this->appends)
-        );
-    }
-
-    /**
-     * Get an attribute array of all arrayable values.
-     */
-    protected function getArrayableItems(array $values): array
-    {
-        if ($this->getVisible() !== []) {
-            return array_intersect_key($values, array_flip($this->getVisible()));
-        }
-
-        return array_diff_key($values, array_flip($this->getHidden()));
-    }
-
-    /**
-     * Get an attribute from the model.
-     */
-    public function getAttribute(string $key): mixed
-    {
-        if (
-            array_key_exists($key, $this->attributes) ||
-            $this->hasGetMutator($key)
-        ) {
-            return $this->getAttributeValue($key);
-        }
-
-        return $this->getRelationValue($key);
-    }
-
-    /**
-     * Get a plain attribute (not a relationship).
-     */
-    protected function getAttributeValue(string $key): mixed
-    {
-        $value = $this->getAttributeFromArray($key);
-
-        // If the attribute has a get mutator, we will call that then return what
-        // it returns as the value, which is useful for transforming values on
-        // retrieval from the model to a form that is more useful for usage.
-        if ($this->hasGetMutator($key)) {
-            return $this->mutateAttribute($key, $value);
-        }
-
-        // If the attribute exists within the cast array, we will convert it to
-        // an appropriate native PHP type dependant upon the associated value
-        // given with the key in the pair. Dayle made this comment line up.
-        if ($this->hasCast($key)) {
-            $value = $this->castAttribute($key, $value);
-        }
-
-        return $value;
-    }
-
-    /**
-     * Get an attribute from the $attributes array.
-     */
-    protected function getAttributeFromArray(string $key)
-    {
-        if (array_key_exists($key, $this->attributes)) {
-            return $this->attributes[$key];
-        }
-
-        return null;
-    }
-
-    /**
-     * Determine if a get mutator exists for an attribute.
-     */
-    public function hasGetMutator(string $key): bool
-    {
-        return method_exists($this, 'get'.Str::studly($key).'Attribute');
-    }
-
-    /**
-     * Get the value of an attribute using its mutator.
-     */
-    protected function mutateAttribute(string $key, mixed $value): mixed
-    {
-        return $this->{'get'.Str::studly($key).'Attribute'}($value);
-    }
-
-    /**
-     * Get the value of an attribute using its mutator for array conversion.
-     */
-    protected function mutateAttributeForArray(string $key, mixed $value): mixed
-    {
-        $value = $this->mutateAttribute($key, $value);
-
-        return $value instanceof Arrayable ? $value->toArray() : $value;
-    }
-
-    /**
-     * Determine whether an attribute should be casted to a native type.
-     */
-    protected function hasCast(string $key): bool
-    {
-        return array_key_exists($key, $this->casts);
-    }
-
-    /**
-     * Determine whether a value is JSON castable for inbound manipulation.
-     */
-    protected function isJsonCastable(string $key): bool
-    {
-        return $this->hasCast($key) &&
-        in_array($this->getCastType($key), ['array', 'json', 'object'], true);
-    }
-
-    /**
-     * Get the type of cast for a model attribute.
-     */
-    protected function getCastType(string $key): string
-    {
-        return trim(strtolower((string) $this->casts[$key]));
-    }
-
-    /**
-     * Cast an attribute to a native PHP type.
-     */
-    protected function castAttribute(string $key, mixed $value): mixed
-    {
-        if (is_null($value)) {
-            return $value;
-        }
-
-        return match ($this->getCastType($key)) {
-            'int', 'integer' => (int) $value,
-            'real', 'float', 'double' => (float) $value,
-            'string' => (string) $value,
-            'bool', 'boolean' => (bool) $value,
-            'object' => $this->fromJson($value, true),
-            'array', 'json' => $this->fromJson($value),
-            'collection' => new BaseCollection($this->fromJson($value)),
-            'datetime' => $this->asDateTime($value),
-            'date' => $this->asDateTime($value)->toDateString(),
-            default => $value,
-        };
-    }
-
-    /**
-     * Set a given attribute on the model.
-     */
-    public function setAttribute(string $key, mixed $value): static
-    {
-        // First we will check for the presence of a mutator for the set operation
-        // which simply lets the developers tweak the attribute as it is set on
-        // the model, such as "json_encoding" an listing of data for storage.
-        if ($this->hasSetMutator($key)) {
-            $method = 'set'.Str::studly($key).'Attribute';
-
-            return $this->{$method}($value);
-        }
-
-        if ($this->isJsonCastable($key) && ! is_null($value)) {
-            $value = $this->asJson($value);
-        }
-
-        $this->attributes[$key] = $value;
-
-        return $this;
-    }
-
-    /**
-     * Determine if a set mutator exists for an attribute.
-     */
-    public function hasSetMutator(string $key): bool
-    {
-        return method_exists($this, 'set'.Str::studly($key).'Attribute');
-    }
-
-    /**
-     * Encode the given value as JSON.
-     */
-    public function asJson(mixed $value): string
-    {
-        return json_encode($value);
-    }
-
-    /**
-     * Return a timestamp as DateTime object.
-     */
-    public function asDateTime(mixed $value): Carbon
-    {
-        return match (true) {
-            $value instanceof Carbon => $value,
-            $value instanceof DateTimeInterface => new Carbon(
-                $value->format('Y-m-d H:i:s.u'),
-                $value->getTimezone()
-            ),
-            is_numeric($value) => Carbon::createFromTimestamp($value),
-            $this->isStandardDateFormat($value) => Carbon::createFromFormat('Y-m-d', $value)->startOfDay(),
-            default => new Carbon($value),
-        };
-    }
-
-    public function getClassName(): string
-    {
-        return (new \ReflectionClass($this))->getShortName();
-    }
-
-    /**
-     * Determine if the given value is a standard date format.
-     */
-    protected function isStandardDateFormat(string $value): bool
-    {
-        return preg_match('/^(\d{4})-(\d{1,2})-(\d{1,2})$/', $value);
-    }
-
-    /**
-     * Decode the given JSON back into an array or object.
-     */
-    public function fromJson(string $value, bool $asObject = false): mixed
-    {
-        return json_decode($value, ! $asObject);
-    }
-
-    /**
-     * Clone the model into a new, non-existing instance.
-     */
-    public function replicate(mixed $except = null): static
-    {
-        $except = $except ?: [];
-
-        $attributes = Arr::except($this->attributes, $except);
-
-        return with(new static())->fill($attributes);
-    }
-
-    /**
-     * Get all of the current attributes on the model.
-     */
-    public function getAttributes(): array
-    {
-        return $this->attributes;
-    }
-
-    /**
-     * Get the mutated attributes for a given instance.
-     */
-    public function getMutatedAttributes(): array
-    {
-        $class = static::class;
-
-        if (! isset(static::$mutatorCache[$class])) {
-            static::cacheMutatedAttributes($class);
-        }
-
-        return static::$mutatorCache[$class];
-    }
-
-    /**
-     * Extract and cache all the mutated attributes of a class.
-     */
-    public static function cacheMutatedAttributes(string $class): void
-    {
-        $mutatedAttributes = [];
-
-        // Here we will extract all of the mutated attributes so that we can quickly
-        // spin through them after we export models to their array form, which we
-        // need to be fast. This'll let us know the attributes that can mutate.
-        if (preg_match_all('/(?<=^|;)get([^;]+?)Attribute(;|$)/', implode(';', get_class_methods($class)), $matches)) {
-            foreach ($matches[1] as $match) {
-                if (static::$snakeAttributes) {
-                    $match = Str::snake($match);
-                }
-
-                $mutatedAttributes[] = lcfirst((string) $match);
-            }
-        }
-
-        static::$mutatorCache[$class] = $mutatedAttributes;
-    }
-
-    /**
-     * Create a new Eloquent Collection instance.
-     *
-     * @param  array  $models
-     * @return \Illuminate\Database\Eloquent\Collection
-     */
-    public function newCollection($models = []): BaseCollection
-    {
-        if ($models instanceof BaseCollection) {
-            return $models;
-        }
-
-        return new BaseCollection($models);
-    }
-
-    /**
-     * Dynamically retrieve attributes on the model.
-     */
-    public function __get(string $key): mixed
-    {
-        $value = null;
-        if ($key !== 'id' && method_exists($this, 'getAugmentedModel') && $this->getAugmentedModel()) {
-            $value = $this->getAugmentedModel()->{$key};
-        }
-
-        return $value ?? $this->getAttribute($key);
-    }
-
-    /**
-     * Dynamically set attributes on the model.
-     */
-    public function __set(string $key, mixed $value): void
-    {
-        $this->setAttribute($key, $value);
-    }
-
-    /**
-     * Determine if the given attribute exists.
-     */
-    public function offsetExists(mixed $offset): bool
-    {
-        return isset($this->{$offset});
-    }
-
-    /**
-     * Get the value for a given offset.
-     */
-    public function offsetGet(mixed $offset): mixed
-    {
-        return $this->{$offset};
-    }
-
-    /**
-     * Set the value for a given offset.
-     */
-    public function offsetSet(mixed $offset, mixed $value): void
-    {
-        $this->{$offset} = $value;
-    }
-
-    /**
-     * Unset the value for a given offset.
-     */
-    public function offsetUnset(mixed $offset): void
-    {
-        unset($this->{$offset});
-    }
-
-    /**
-     * Get the endpoint
-     */
-    public function getEndpoint($type): string
-    {
-        return $this->endpoints[$type];
-    }
-
-    /**
-     * Implement Basic URLRoutable functions
-     */
-    public function getRouteKey()
-    {
-        return $this->getAttribute($this->getRouteKeyName());
-    }
-
-    public function getRouteKeyName()
-    {
-        return $this->getKeyName();
-    }
-
-    public function resolveRouteBinding($value, $field = null)
-    {
-        return $this->entity;
-    }
-
-    public function resolveChildRouteBinding($childType, $value, $field)
-    {
-    }
-
-    public function getKeyName()
-    {
-        return $this->primaryKey;
-    }
-
-    /**
-     * Determine if an attribute exists on the model.
-     */
-    public function __isset(string $key): bool
-    {
-        return (isset($this->attributes[$key]) || isset($this->relations[$key])) ||
-            ($this->hasGetMutator($key) && ! is_null($this->getAttributeValue($key)));
-    }
-
-    /**
-     * Unset an attribute on the model.
-     */
-    public function __unset(string $key): void
-    {
-        unset($this->attributes[$key]);
-    }
-
-    /**
-     * Handle dynamic static method calls into the method.
-     */
-    public static function __callStatic(string $method, array $parameters): mixed
-    {
-        $instance = new static();
-
-        return call_user_func_array([$instance, $method], $parameters);
-    }
-
-    /**
-     * Convert the model to its string representation.
-     */
-    public function __toString(): string
-    {
-        return $this->toJson();
-    }
-}
